Added biicode support.
Unconventional to a regular biicode block, we are not specifying the sources required to build this library.
Instead, we just link it to the CppMicroServices library.

The "core/include/usConfig.h" file is present as this allows bii to resolve a block's required headers.
Inclusion of generated headers directly results in a missing dependency during compilation.