Fix Windows static destruction corner case
It's been observed that when CppMicroServices is used via JNI, during process shutdown, even if the Framework was explicitly shut down, a crash can still occur.  The fix is to remove unnecessary access to thread and synchronizxation objects unless required.  

This also includes a minor fix to resolve a GCC build warning.

Signed-off-by: The MathWorks, Inc. Roy.Lurie@mathworks.com