Adds tests for ResourceCompiler
Primarily, this commit adds tests for ResourceCompiler.
Refer to https://github.com/CppMicroServices/CppMicroServices/issues/162
The following use cases are covered:
1. Create a zip file with just the manifest
2. Create a zip file with manifest and an example resource
3. Create a zip file with an example resource
4. Merge the zip file created in 3. and add a manifest/resource
to a new zip file.
5. Append a zip file to a sample bundle.

For this purpose, a new file ResourceCompilerTest.cpp is created.
The new test is added an entry to the driver program
usFrameworkTestDriver.
Also, a few cross-platform filesystem functions are added to TestUtils.

Signed-off-by: The Mathworks Inc <Roy.Lurie@mathworks.com>