Fixes ResourceCompiler based on review feedback
Addresses
https://github.com/CppMicroServices/CppMicroServices/issues/162

Summary:
1. ZipFile moved to a separate file and made more container-like.

2. Made the tests use absolute paths instead of relative paths (except
for tests testing --res-add option)

3. Improved exception testing. Now, the test checks if an exception is
thrown in any of the creation/deletion of the temporary files, exits the
test, if so.  Similarly for the command line tests themselves.
  Added a test macro US_TEST_NO_EXCEPTION_OTHERWISE_RETURN, which not
only tests that an exception is not thrown but if it's thrown, exits the
test. This is useful because if certain operation (like creation of the
sample directory hierarchy) doesn't work, none of the tests are going to
work, so the right approach at that point is to exit the test.

4. Added a CheckFileAndRemove utility function that removes a file only
if it exists.

5. Removed GetExecutablePath() utility function in favor of
testing::BIN_PATH.

6. Added a few more tests to exercise more options of the
ResourceCompiler. Also tested a variety of failure modes.

7. Changed relative path invocation to absolute path wherever possible

8. Changes based on build failure in gcc 4.9.2 and mac test failures.

Minor improvements in ResourceCompiler documentation.

Addresses issue 5 of
https://github.com/CppMicroServices/CppMicroServices/issues/161

1. fixed an issue where the invocation with only "--help" option
would throw a "At least one of the options (--bundle-file | --out-file)
is required" error.

2. removed usage information displayed along with errors.

Signed-off-by: The Mathworks Inc <Roy.Lurie@mathworks.com>