Robustify ResourceCompiler implementation
Cleaned up documentation and added some notes on the constraints of
various options.

Ensured that only res-add and manifest-add can be specified multiple
times. Issue 4 in
https://github.com/CppMicroServices/CppMicroServices/issues/161

Ensured bundle-name option is provided if res-add or manifest-add is
provided. Added a warning that bundle-name option is unnecessary if
neither res-add nor manifest-add is provided. Issue 7 and 8 of #161.

Removed code comment that a bundle can contain only one zip blob.
Issue 6 in #161.

Signed-off-by: The Mathworks Inc <Roy.Lurie@mathworks.com>