Tests for resource compiler (#169)
* Adds tests for ResourceCompiler

Primarily, this commit adds tests for ResourceCompiler.
Refer to https://github.com/CppMicroServices/CppMicroServices/issues/162
The following use cases are covered:
1. Create a zip file with just the manifest
2. Create a zip file with manifest and an example resource
3. Create a zip file with an example resource
4. Merge the zip file created in 3. and add a manifest/resource
to a new zip file.
5. Append a zip file to a sample bundle.

For this purpose, a new file ResourceCompilerTest.cpp is created.
The new test is added an entry to the driver program
usFrameworkTestDriver.
Also, a few cross-platform filesystem functions are added to TestUtils.

Signed-off-by: The Mathworks Inc <Roy.Lurie@mathworks.com>

* Fixes ResourceCompiler based on review feedback

Addresses
https://github.com/CppMicroServices/CppMicroServices/issues/162

Summary:
1. ZipFile moved to a separate file and made more container-like.

2. Made the tests use absolute paths instead of relative paths (except
for tests testing --res-add option)

3. Improved exception testing. Now, the test checks if an exception is
thrown in any of the creation/deletion of the temporary files, exits the
test, if so.  Similarly for the command line tests themselves.
  Added a test macro US_TEST_NO_EXCEPTION_OTHERWISE_RETURN, which not
only tests that an exception is not thrown but if it's thrown, exits the
test. This is useful because if certain operation (like creation of the
sample directory hierarchy) doesn't work, none of the tests are going to
work, so the right approach at that point is to exit the test.

4. Added a CheckFileAndRemove utility function that removes a file only
if it exists.

5. Removed GetExecutablePath() utility function in favor of
testing::BIN_PATH.

6. Added a few more tests to exercise more options of the
ResourceCompiler. Also tested a variety of failure modes.

7. Changed relative path invocation to absolute path wherever possible

8. Changes based on build failure in gcc 4.9.2 and mac test failures.

Minor improvements in ResourceCompiler documentation.

Addresses issue 5 of
https://github.com/CppMicroServices/CppMicroServices/issues/161

1. fixed an issue where the invocation with only "--help" option
would throw a "At least one of the options (--bundle-file | --out-file)
is required" error.

2. removed usage information displayed along with errors.

Signed-off-by: The Mathworks Inc <Roy.Lurie@mathworks.com>

* Robustify ResourceCompiler implementation

Cleaned up documentation and added some notes on the constraints of
various options.

Ensured that only res-add and manifest-add can be specified multiple
times. Issue 4 in
https://github.com/CppMicroServices/CppMicroServices/issues/161

Ensured bundle-name option is provided if res-add or manifest-add is
provided. Added a warning that bundle-name option is unnecessary if
neither res-add nor manifest-add is provided. Issue 7 and 8 of #161.

Removed code comment that a bundle can contain only one zip blob.
Issue 6 in #161.

Signed-off-by: The Mathworks Inc <Roy.Lurie@mathworks.com>

* Modify the test based on review feedback

Specifically,
1. Changed the names of functions to be more descriptive

2. Added a few more failure modes tests

3. Changed name US_TEST_NO_EXCEPTION_OTHERWISE_RETURN to
US_TEST_NO_EXCEPTION_REQUIRED. Changed the US_TEST_NO_EXCEPTION_REQUIRED
macro to use US_TEST_FAILED_MSG

* Fix test based on review

In summary, check for the existence of individual entries rather than the index
so that the test is independent of the index. To achieve this, the method
getNames() is added to ZipFile class which returns a vector of all names. Made
other minor improvements.

Signed-off-by: The Mathworks Inc <Roy.Lurie@mathworks.com>

* Refactor error checking

Essentially, factored out the command line error checking code into a function
for clarity. Also, improved the lambda function which checks for existence of
multiple same arguments.

Signed-off-by: The Mathworks Inc <Roy.Lurie@mathworks.com>

* Check command line invocation return code

Added tests which check that the Resource Compiler command line invocations
(for the typical use cases) return zero.

Signed-off-by: The Mathworks Inc <Roy.Lurie@mathworks.com>

* Fix build error

Fixed miniz.c path

* Fix Travis test failures

Attempts to fix #171 by bracketing the command line string with single quotes.

* Fix Travis test failure

Attempts to fix #171 by bracketing the command line string with double
quotes.

* Fix test failures in Travis

This addresses #171

Introduced a function escapePath() that transform the paths to contain escape
sequence before spaces and parenthesis. This should hopefully fix the test
failures in Travis due to the path containing spaces/parenthesis.

Made a few more minor improvements.

Signed-off-by: The Mathworks Inc <Roy.Lurie@mathworks.com>

* Fix a minor issue

The path escaping for #171 has to happen before checking the existence
of the executable.

* Fix Travis CI test failures

Addresses #171.

This fix escapes the spaces and parenthesis with a single backslash
instead of the erroneous double backslash in the previous commit.
Hopefully this fixes the Travis test issues.