Support multiple listeners.
Fix based on Sascha's PR review.

Modify bundle and framework listeners to be an unordered_map with
ListenerTokenIds as keys. This improves the removal of these types of listeners
to O(1).

Deprecate addition of listeners functions which take member functions of
objects. Clients can use std::bind together with the other Add*Listener
functions to achieve the same result.

Modify RemoveListener call to take token by value. With this change, clients
have to explicitly move the tokens.