reduce the # of open file handles
Installing thousands of bundles within a process can cause undefined behavior when the OS open file limit has been reached. 

If the bundle only contains a manifest.json, close the file after reading the manifest.json contents.
If the bundle meta-data is read again, open the file and leave it open until the bundle is uninstalled.

Signed-off-by: The Mathworks Inc <Roy.Lurie@mathworks.com>