change logic around US_LIB_POSTFIX.
Changed the setting of the value in GlobalConfig.h.in to be selected by #ifdef NDEBUG so
that if the type of the build is changed from Debug to Release (or whatever) without
re-running cmake, the right thing will happen. The way it was, the value was set at cmake
time, and then if the type of build changed the value of US_LIB_POSTFIX did not change to
reflect the change in build type.