Fix a deadlock when using a thread pool with one thread [ci skip]
When using one thread to do state transitions a deadlock can occur whereby a state transition happening on the one thread is waiting for another transition to occur first. A solution is to guard the state change and the posting of work to the thread so that they can't be interleaved.
[ci skip]
Signed-off-by: The MathWorks, Inc. <jdicleme@mathworks.com>