Updates from code review
Updated the prototype code for integrating ConfigAdmin into DS with the changes from the code review. The most important changes were:
- Moved static map of listeners out of ConfigurationListener into a new class ConfigurationNotifier and made it non-static.
- Added mutex to ConfigurationManager class to synchronize access to the configProperties and mergedProperties members. This mutex prevents collisions between simultaneous configuration object updates.
- Added asynchonous task to ComponentConfigurationImpl::ConfigChangedState so that the important part of the functionality in there (Sending modified properties to the component instance, registering a component or deactivating a component) happens on a new thread instead of on the ConfigurationListener thread.
Note to reviewers. Moving the listeners map to a new class changed a lot of file. Only one object of the class is constructed by SCRActivator and then passe along to the ComponentConfigurationImpl constructor. A lot of constructors along the way were affected and a lot of tests that create ComponentConfigurationImpl objects were affected. Most of the changes to the files were small and related to that change. The significant changes which need to be reviewed are in ComponentConfigurationImpl.cpp, ConfigurationManager.cpp, ConfigurationNotifier.cpp and ConfigurationListener.cpp Signed-off-by: The MathWorks, Inc. <pelliott@mathworks.com>