Removed unecessary wait loops from DS and ConfigAdmin tests.
DS and ConfigAdmin tests that were calling ConfigAdmin to update or remove configuration objects used to have to loop while waiting for the asynchronous operation to complete. The Update and Remove operations have been changed to not return to the caller until after the asynchronous operation is complete so looping is no longer necessary. Removed the unecessary looping code from the DS and ConfigAdmin tests. Signed-off-by: The MathWorks, Inc. <pelliott@mathworks.com>