Fixed TestComponentConfigurationImpl.VerifyConcurrentActivateDeactivate
Fixed sporadic failure in Fixed TestComponentConfigurationImpl.VerifyConcurrentActivateDeactivate. The problem was caused because the test was destroying the MockComponentConfigurationImpl before the last Activate call was finished. An exception was thrown.  Changed the code so it will wait for the last call to be completed before destroying the object. Signed-off-by: The MathWorks, Inc. <pelliott@mathworks.com>