Update
ConfigurationAdmin notifies DS when a configuration object is updated. The action taken depends on the configuration-policy, whether the component was satisfied before the update, whether the component is still satisfied after the update, whether the service author has provided a Modified method and the current state of the component. Signed-off-by: The MathWorks, Inc <pelliott@mathworks.com>