Fix TestConcurrentBindUnbind error
An exception was being thrown when running the TestConcurrentBindUnbind method because DS was in the middle of binding to a service reference while another thread was unregistering the service. The sObjs.GetService call in ComponentContextImpl::AddToBoundServicesCache was throwing the exception because the ServiceReferenceBasePrivate object had been cleared out by the unregister operation and was returning an empty InterfaceMap. Changed ServiceObjects.cpp to check for the return of an empty map and handle the condition. Signed-off-by: The MathWorks, Inc. <pelliott@mathworks.com>