Moved DoesModifiedMethodExist to ComponentInstance public interface
Moved DoesModifiedMethodExist to ComponentInstance public interface from ComponentContextImpl. The previous implementation with the method in ComponentContextImpl didn't work on linux and mac. In that implementation the ComponentInstanceImpl.hpp file included the ComponentContextImpl.hpp file which was not a public interface so it couldn't be found. Signed-off-by: The MathWorks, Inc.  <pelliott@mathworks.com>