Fix ConfigurationAdmin ListConfigurations
There was an order dependency for services requiring configuration objects. If the configuration object was available before the component was initialized, DS would not realize the configuration object was available and would not satisfy the component. The problem was with the ListConfigurations method. It was not returning the configuration objects correctly even though they existed in the Configuration Admin repository. Fixed the ListConfigurations method and added two tests. Signed-off-by <pelliott@mathworks.com>