Fix Segfault happening in TestUpdateConfiguration
This segfault was caused by a nullptr being returned for a GetInstance call following GetComponentConfigurationDTOs call that indicated that the component state was SATISFIED. The problem was that the state was SATISFED but because of a race condition the component wasn't yet registered so GetInstance was returning a nullptr. Fixed the GetComponentConfigurationDTOs call to wait until the state transition is complete before returning the state. Signed-off-by <pelliott@mathworks.com>