Resolved race condition in ConfigAdminTests. testServiceUpdated
There was a race condition in ConfigAdminTests.testServiceUpdate caused by the fact that AddingService was trying to add the cm.testservice configuration object and AddConfigurations was also trying to add it. AddingService was adding a blank configuration object and calling the ManagedService interface's Updated method. AddConfigurations was adding the correct configuration object (found in the manifest.json file) and calling the ManagedService interface's Updated method. According to OSGI the addition of a blank configuration object to the repository is not supposed to result in an Update notification to the ManagedService so the solution to the race condition was to stop sending the notification when AddingService added the blank configuration object. Signed-off-by <pelliott@mathworks.com>