Configuration Admin fix for duplicate Update Notifications
According to OSGI, Update notifications are only supposed to be sent by ConfigurationAdmin when a configuration object has been updated. A configuration object which has not yet been Updated should not result in an Update Notification. Also, a Removed Notification may only be sent if an Updated notification was previously sent. There were several places in our ConfigurationAdmin code where Updated and Removed notifications were being sent to services in error. This change fixes this.

Each Configuration object contains a changeCount. The changeCount is set equal to 1 when the Configuration object is created and incremented each time the Configuration object is Updated. Updated and Removed notifications will only be sent for Configuration objects with a changeCount > 1. The AddConfigurations method, adds Configuration objects for configurations read from the manifest.json file. These Configuration objects are added to the repository with changeCount = 2 since the construction of these objects is equivalent to a creation and an update.

The Configuration object now has a method called IsUpdated which returns true for Configuration objects with a changeCount > 1 and false otherwise.

ConfigurationAdminImpl::NotifyConfigurationUpdated now accepts a second input parameter called removeUpdate. This is true when the Configuration object has been removed but IsUpdated was true before the object was removed. NotifyConfigurationUpdate is used for both Updated and Removed operations. For a Removed operation the Configuration object has already been removed and IsUpdated cannot be called so this input parameter controls whether or not a Removed notification will be sent.

ConfigurationAdmin has service trackers that monitor services that implement the ManagedService and ManagedServiceFactory interfaces. It receives an AddingService notification when one of these services is registered. ConfigurationAdmin is responsible for notifying services of Configuration objects that were added to the repository (and updated) prior to the service being registered. There was a problem with our code because the configuration repository was being changed while the AddingService methods were accessing the repository. The race condition resulted in two Updated notifications being sent for the same Update. This change fixes this problem by locking the configuration repository at the beginning of the AddingService method so that no changes can occur while the AddingService method is in process.
Signed-off-by The MathWorks, Inc. <pelliott@mathworks.com>