BugFix when creating instance name for factory components
The auto generation code uses the component name from the manifest.json file when creating the name of the constructor and destructor for the service instance. The factory component was using the name of the implementation class. This coding change fixes this bug so the name of the component is always used. The name of the component is an optional parameter in the manifest.json file. It defaults to the implementation class when not provided. The implementation class will be used for the instance name in those cases. Signed off by The MathWorks, Inc. <pelliott@mathworks.com>