Fix broken static build configurations on macOS (#774)
* Fix broken static build configurations on macOS

Fixes #738

Removed the ability to append ZIP files to the end of archives files for all supported OS platforms. Specifically the clang linker no longer tolerates linking archive files with arbitrary data appended to the end.

Signed-off-by: The MathWorks, Inc. <jdicleme@mathworks.com>

* Fix examples tests

the tutorial example has tests which use an installed CppMicroServices SDK to test whether the example code can be compiled using the SDK. In this case, the CppMicroServices target property doesn't have the path to the metadata zip set.

To resolve this issue:
1. install the metadata zip files alongside the static libraries for all CppMicroServices components only if building static libraries.
2. Set this property for cases where the CppMicroServices SDK is used by downstream clients.

Signed-off-by: The MathWorks, Inc. <jdicleme@mathworks.com>

Signed-off-by: The MathWorks, Inc. <jdicleme@mathworks.com>
Co-authored-by: Jeff <DiClemente>