Fix race with DS service object construction
A race occurs when DS is constructing a service object with injectable mandatory service dependencies whereby any bound service dependencies could no longer be available because another thread unregistered any of the bound service dependencies.

Given that a ComponentContext is constructed only after all the service dependencies has been satisfied, if during ComponetContext construction any of the bound service dependencies are no nullptrs then throw and fail to construct the service object. This fixes a failure mode where a service constructor can be called with a nullptr service object when the service dependency is defined as mandatory and injection is true.

Also noticed that one of the test bundle project was exceeding the path limit on Windows, so shortened the project name to mitigate that problem.

Signed-off-by: The MathWorks, Inc. <jdicleme@mathworks.com>