Added a test case
Test case shows that top level flat keys are preferred over nested keys if both are present pointing
to an equivalent value. For instance, given the following json:

{
  "a.b" : 5,
  "a" : {
    "b" : 10
  }
}

The query for "(a.b=5)" will match, and "(a.b=10)" will not match.

Also updated the doxygen documentation for LDAPFilter to explain this extension to the behavior of
LDAP filters.