ComponentManagermemory reallocation bug fix (#834)
* ComponentManagermemory reallocation bug fix

The ComponentManagerImpl object contained a shared_ptr to the vector of all ComponentManagerImpl objects. When adding an item to this vector, the resizing of the vector caused a memory allocation error. Fixed by removing the vector from the ComponentManagerImpl object.

* Fix for ConfigurationNotifier::CreateFactoryComponent memory allocation error

When a ComponentManager is created, it must be added to the managers map in SCRBundleExtension. This is necessary so that when the bundle is stopped, the ComponentManager can be destroyed.
The old code was holding a shared_ptr to the managers map in the ComponentManagerImpl object. This caused a memory reallocation error when a ComponentManager was added to the map and the map needed to be resized.
This fix adds an AddComponentManager method to the SCRBundleExtension object so that CreateFactoryComponent can add a manager to the managers map when it is created. In order to find the SCRBundleExtension, the map of SCRBundleExtensions was moved out of SCRActivator and into a new class called SCRExtensionRegistry.

* Rest of the ConfigurationNotifier fix

Signed-off-by: The MathWorks, Inc. <pelliott@mathworks.com>

* Fix errors in minimum gcc build

Signed-off by: The MathWorks Inc. < pelliott@mathworks.com>

* Responded to code review comments

Responded to code review comments. Added tests for SCRBundleExtension and SCRExtensionRegistry. Signed off by - The MathWorks, Inc. <pelliott@mathworks.com>

* Update ConfigurationNotifier.cpp

Update error message. Signed off by The MathWorks, Inc. <pelliott@mathworks.com>

* Update SCRExtensionRegistry.hpp

Formatting issue. signed-off by The MathWorks, Inc. <pelliott@mathworks.com>

* CreateFactoryComponent changes

ConfigurationNotifier::CreateFactoryComponent will log an exception if it can't find the SCRBundleExtension in the ExtensionRegistry instead of throwing a std::runtime_error exception. Signed-off-by: The MathWorks, Inc. <pelliott@mathworks.com>

* Update ConfigurationNotifier.cpp

Fixed call to std::exception. Signed-off-by The MathWorks, Inc. <pelliott@mathworks.com>

* Update ConfigurationNotifier.cpp

Removed throw in CreateFactoryComponent. Error message is already logged, no throw needed. Signed-off-by - The MathWorks, Inc. <pelliott@mathworks.com>

---------

Signed-off-by: The MathWorks, Inc. <pelliott@mathworks.com>

Adapt implementation and test code to C++14:
* move initialization statements out of if clause
* adapt assignment construction in test code calling deleted move constructor

Signed-off-by: Ingmar Sittl <ingmar.sittl@elektrobit.com>