responded to comments and refactored some code.
Refactored code is around the use of the option::Option array. Instead of a std::unique_ptr to an
allocated array of option::Option object, use a std::vector and pass a reference to the vector to
checkSanity(). This should avoid warnings about pointer arithmetic and is also more appropriate code
for the use case.