I0184 (#1102)
* Issue #184: added check for bundle name matching manifest value

Causes test failures... not exactly sure how to resolve them.

* Updated test cases to work properly with new restrictions

* added test for mismatched bundle name and bundle.symbolic_name

* responded to comments and refactored some code.

Refactored code is around the use of the option::Option array. Instead of a std::unique_ptr to an
allocated array of option::Option object, use a std::vector and pass a reference to the vector to
checkSanity(). This should avoid warnings about pointer arithmetic and is also more appropriate code
for the use case.

---------

Co-authored-by: Michael Carney <carney@macmini.mynetworksettings.com>