Namespace renaming tool (#1050)
Following changes are to introduce tool to rename namespace of CppMicroServices codebase.
Tool: change_namespace
Usage: change_namespace --cppms= --namespace=<new_namespace_word> --namespace-alias (if we want to alias to the original namespace) <output_directory>

This tool will copy the all the files with namespace "cppmicroservices" replaced with the new word provided in the output directory.

Signed-off-by: The MathWorks, Inc. mphadnis@mathworks.com